Changed user id type from Integer to Long

+++ src/main/java/com/rep/api/user/UserRepository.java
@@ -1,11 +1,10 @@
 package com.rep.api.user;
 
-import com.rep.api.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
 }
